Made updates to HTML, CSS, and JS files
Updated HTML file title to reflect app functionality more accurately (building out a forecast feature will come later). Added a conditional if statement of the hours in the time element. Removed the emoji and created a new div contained with the id of "icon." Restyled elements to reposition them aesthetically. Added two lines, where the first line is selecting the icon element that has nothing in it and then injecting the image
using the URL based on the API.